remove unnecessary function to abstract class
